chore: Move bootstrap package to runtime
